fixing warning, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@14175 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/geom/PositionVector.cpp
@@ -987,7 +987,7 @@ PositionVector::move2side(SUMOReal amount) {
 
 Line
 PositionVector::lineAt(int pos) const {
-    assert(size() > pos + 1);
+    assert((int)size() > pos + 1);
     return Line((*this)[pos], (*this)[pos + 1]);
 }
